backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.1.2,<5.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# File Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# PDF Generation
reportlab==4.0.8

# Utilities
python-dateutil==2.8.2

# Development
pytest==7.4.4
httpx==0.26.0
